+++ src/pages/order/allOrder/components/UpdatePaymentStatus.jsx
@@ -1,15 +1,30 @@
-import React from "react";
+import { Loader2 } from "lucide-react";
+import React, { useState } from "react";
 
-function UpdatePaymentStatus({
-  handlePaymentStatusUpdate,
-  selectedOrder,
-  paymentStatusLoading,
-}) {
+function UpdatePaymentStatus({ handlePaymentStatusUpdate, selectedOrder }) {
+  const [updatingStatus, setUpdatingStatus] = useState(null);
+
   if (!selectedOrder) return null;
+
+  const statuses = [
+    { key: "PENDING", label: "অপেক্ষমান", color: "blue" },
+    { key: "PAID", label: "পরিশোধিত", color: "green" },
+    { key: "FAILED", label: "ব্যর্থ", color: "red" },
+    { key: "REFUNDED", label: "ফেরত", color: "gray" },
+  ];
 
   const current = (selectedOrder.paymentStatus ?? "PENDING")
     .toString()
     .toUpperCase();
+
+  const handleClick = async (statusKey) => {
+    setUpdatingStatus(statusKey);
+    try {
+      await handlePaymentStatusUpdate(selectedOrder.id, statusKey);
+    } finally {
+      setUpdatingStatus(null);
+    }
+  };
 
   return (
     <div className="border-t border-gray-200 pt-4 mb-6">
@@ -18,55 +33,26 @@ function UpdatePaymentStatus({
       </h4>
 
       <div className="flex flex-wrap gap-2">
-        <button
-          onClick={() => handlePaymentStatusUpdate(selectedOrder.id, "PENDING")}
-          disabled={current === "PENDING" || paymentStatusLoading}
-          className={`px-3 py-2 rounded-md text-sm ${
-            current === "PENDING"
-              ? "bg-blue-100 text-blue-800 cursor-not-allowed"
-              : "border border-blue-300 text-blue-700 hover:bg-blue-50"
-          }`}
-        >
-          অপেক্ষমান
-        </button>
+        {statuses.map((status) => {
+          const isActive = current === status.key;
+          const isLoading = updatingStatus === status.key;
 
-        <button
-          onClick={() => handlePaymentStatusUpdate(selectedOrder.id, "PAID")}
-          disabled={current === "PAID" || paymentStatusLoading}
-          className={`px-3 py-2 rounded-md text-sm ${
-            current === "PAID"
-              ? "bg-green-100 text-green-800 cursor-not-allowed"
-              : "border border-green-300 text-green-700 hover:bg-green-50"
-          }`}
-        >
-          পরিশোধিত
-        </button>
-
-        <button
-          onClick={() => handlePaymentStatusUpdate(selectedOrder.id, "FAILED")}
-          disabled={current === "FAILED" || paymentStatusLoading}
-          className={`px-3 py-2 rounded-md text-sm ${
-            current === "FAILED"
-              ? "bg-red-100 text-red-800 cursor-not-allowed"
-              : "border border-red-300 text-red-700 hover:bg-red-50"
-          }`}
-        >
-          ব্যর্থ
-        </button>
-
-        <button
-          onClick={() =>
-            handlePaymentStatusUpdate(selectedOrder.id, "REFUNDED")
-          }
-          disabled={current === "REFUNDED" || paymentStatusLoading}
-          className={`px-3 py-2 rounded-md text-sm ${
-            current === "REFUNDED"
-              ? "bg-gray-200 text-gray-800 cursor-not-allowed"
-              : "border border-gray-300 text-gray-700 hover:bg-gray-50"
-          }`}
-        >
-          ফেরত
-        </button>
+          return (
+            <button
+              key={status.key}
+              onClick={() => handleClick(status.key)}
+              disabled={isActive || isLoading}
+              className={`px-3 py-2 rounded-md text-sm flex items-center justify-center gap-1 ${
+                isActive
+                  ? `bg-${status.color}-100 text-${status.color}-800 cursor-not-allowed`
+                  : `border border-${status.color}-300 text-${status.color}-700 hover:bg-${status.color}-50`
+              }`}
+            >
+              {isLoading && <Loader2 className="w-4 h-4 animate-spin" />}
+              {status.label}
+            </button>
+          );
+        })}
       </div>
     </div>
   );
