+++ src/pages/order/utils/formatDate.js
@@ -1,0 +1,8 @@
+export const formatDate = (dateString) =>
+  new Date(dateString).toLocaleString("bn-BD", {
+    year: "numeric",
+    month: "short",
+    day: "numeric",
+    hour: "numeric",
+    minute: "numeric",
+  });
